1365 handles create package under handles and move handles utilities into it (#1373)
# Goal
The goal of this PR is to create a handles-utils package with API for
handles. The executable/binary to generate confusables from a UTF-8 data
file is not automatically built.

Part of #1365
