+++ pallets/handles/Cargo.toml
@@ -16,11 +16,9 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { package = "parity-scale-codec", version = "3.1.5", default-features = false, features = ["derive",] }
 log = { version = "0.4.17", default-features = false }
 scale-info = { version = "2.2.0", default-features = false, features = ["derive",] }
-unicode-normalization = { version = "0.1.22", default-features = false }
 numtoa = "0.2.4"
 rand = { version = "0.8.5", default-features = false, features = ["small_rng"] }
 twox-hash = { version = "1.6.3", default-features = false, features = ["digest_0_10"] }
-phf = { version = "0.11", default-features = false, features = ["macros"] }
 # Substrate
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false, optional = true, branch = "polkadot-v0.9.36" }
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.36" }
@@ -31,7 +29,7 @@ sp-runtime = { git = "https://github.com/paritytech/substrate", default-features
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.36" }
 # Frequency related dependencies
 common-primitives = { default-features = false, path = "../../common/primitives" }
-
+handles-utils = { default-features = false, path = "src/handles-utils"}
 [dev-dependencies]
 env_logger = "0.10.0"
 pretty_assertions = "1"
