+++ pallets/handles/src/handles-utils/src/data/confusable_characters.txt
