+++ pallets/handles/src/handles-utils/src/lib.rs
@@ -1,0 +1,12 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod confusables;
+pub mod converter;
+mod types;
+pub mod validator;
+
+pub use confusables::*;
+pub use converter::*;
+pub use validator::*;
+#[cfg(test)]
+mod tests;
