+++ pallets/handles/src/handles-utils/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod converter_tests;
+mod validator_tests;
