+++ pallets/handles/src/handles-utils/src/validator.rs
@@ -2,7 +2,10 @@
 //!
 //! `handle_validator` provides a `HandleValidator` struct to determine the validity of a given
 //! user handle.
-use sp_std::prelude::Vec;
+
+#![cfg_attr(not(feature = "std"), no_std)]
+extern crate alloc;
+use alloc::vec::Vec;
 
 /// A validator for user handles.
 ///
