+++ pallets/handles/src/tests/mod.rs
@@ -1,6 +1,4 @@
 pub mod mock;
 
-mod converter_tests;
 mod handle_creation_tests;
 mod handle_retirements_tests;
-mod validator_tests;
