--- pallets/handles/src/utils/mod.rs
@@ -1,3 +1,0 @@
-mod confusables;
-pub mod converter;
-pub mod validator;
